--- packages/hono-takibi/src/result/err.test.ts
@@ -1,14 +1,0 @@
-import { describe, it, expect } from 'vitest'
-import { err } from '.'
-
-// Test run
-// pnpm vitest run ./src/result/err.test.ts
-
-describe('err', () => {
-  it.concurrent('should create an Err result', () => {
-    const result = err('err')
-    expect(result).toEqual({ ok: false, error: 'err' })
-    expect(result.ok).toBe(false)
-    expect(result.error).toBe('err')
-  })
-})
