--- packages/hono-takibi/src/result/err.ts
@@ -1,10 +1,0 @@
-import type { Err } from './index.js'
-
-/**
- * Creates an error result.
- * @param error - The error of the result.
- * @returns An Err object containing the error.
- */
-export function err<E>(error: E): Err<E> {
-  return { ok: false, error }
-}
