--- packages/hono-takibi/src/result/ok.ts
@@ -1,10 +1,0 @@
-import type { Ok } from './index.js'
-
-/**
- * Creates a successful result.
- * @param value - The value of the successful result.
- * @returns An Ok object containing the value.
- */
-export function ok<T>(value: T): Ok<T> {
-  return { ok: true, value }
-}
