--- packages/hono-takibi/src/result/types.ts
@@ -1,3 +1,0 @@
-export type Ok<T> = { ok: true; value: T }
-export type Err<E> = { ok: false; error: E }
-export type Result<T, E> = Ok<T> | Err<E>
